refactor(notifications): decompose SchedulerService bodies into Collections/Retention/OwnerReport notifiers
Behavior-preserving (0 $tx/JE, NO existing spec -> verbatim). Facade keeps all 20 @Cron
methods + decorators + reportCronFailure + try/catch shells; the 6 logic-heavy handler
BODIES delegate to 3 internally-constructed sub-services (12 pass-throughs stay inline;
module untouched). Handler ordering + best-effort independent retention writes preserved.
tsc 0, notifications tests green, lint 0.
